chore: adapt new models into actions and repositories
